Backgrounds
Stephen and I have repaired backgrounds such that they function.

A lot of code was restructured, and it ended up being much cleaner.
The background variable is no longer passed in to xGame, but is included in world.

The ImageSpriteGroup was removed and the BackgroundGroup module was added in its place.

A World class was created that is to contain everything that makes an xGame world.

The examples were all updated to reflect this new interface.
Some even have backgrounds.